+++ CumulusMX/Stations/EcowittExtraLogFile .cs
@@ -17,6 +17,8 @@ namespace CumulusMX.Stations
 		private readonly Dictionary<string, int> FieldIndex = [];
 		private readonly int interval;
 
+		public bool HeaderValid { get; }
+
 		public EcowittExtraLogFile(List<string> data, Cumulus cumul, int interval)
 		{
 			cumulus = cumul;
@@ -24,7 +26,7 @@ namespace CumulusMX.Stations
 			this.interval = interval;
 
 			// parse the header
-			HeaderParser(data[0]);
+			HeaderValid = HeaderParser(data[0]);
 		}
 
 		public SortedList<long, EcowittApi.HistoricData> DataParser()
@@ -322,7 +324,7 @@ namespace CumulusMX.Stations
 			return baseRec;
 		}
 
-		private void HeaderParser (string header)
+		private bool HeaderParser (string header)
 		{
 			// Time,          CH1 Temperature(℃),CH1 Dew point(℃),CH1 HeatIndex(℃),CH1 Humidity(%), CH2 Temperature(℃),CH2 Dew point(℃),CH2 HeatIndex(℃),CH2 Humidity(%), CH3 Temperature(℃),CH3 Dew point(℃),CH3 HeatIndex(℃),CH3 Humidity(%),CH4 Temperature(℃), CH4 Dew point(℃),CH4 HeatIndex(℃),CH4 Humidity(%),CH5 Temperature(℃), CH5 Dew point(℃),CH5 HeatIndex(℃),CH5 Humidity(%),CH6 Temperature(℃), CH6 Dew point(℃),CH6 HeatIndex(℃),CH6 Humidity(%),CH7 Temperature(℃),CH7 Dew point(℃),CH7 HeatIndex(℃),CH7 Humidity(%),CH8 Temperature(℃),CH8 Dew point(℃), CH8 HeatIndex(℃),CH8 Humidity(%), WH35 CH1hum(%),WH35 CH2hum(%),WH35 CH3hum(%),WH35 CH4hum(%),WH35 CH5hum(%),WH35 CH6hum(%),WH35 CH7hum(%),WH35 CH8hum(%),             Thunder count,Thunder distance(km),AQIN Temperature(℃),AQIN Humidity(%),AQIN CO2(ppm),AQIN PM2.5(ug/m3),AQIN PM10(ug/m3),AQIN PM1.0(ug/m3),AQIN PM4.0(ug/m3),SoilMoisture CH1(%),                                                      SoilMoisture CH2(%),                                                      SoilMoisture CH3(%),                                                      SoilMoisture CH4(%),                                                      SoilMoisture CH5(%),                                                      SoilMoisture CH6(%),                                                      SoilMoisture CH7(%),                                                      SoilMoisture CH8(%),                                                      SoilMoisture CH9(%),                                                      SoilMoisture CH10(%),                                                         SoilMoisture CH11(%),                                                         SoilMoisture CH12(%),                                                         SoilMoisture CH13(%),                                                         SoilMoisture CH14(%),                                                         SoilMoisture CH15(%),                                                         SoilMoisture CH16(%),                                                         Water CH1,Water CH2,Water CH3,Water CH4,Pm2.5 CH1(ug/m3),Pm2.5 CH2(ug/m3),Pm2.5 CH3(ug/m3),Pm2.5 CH4(ug/m3),WN34 CH1(℃),WN34 CH2(℃),WN34 CH3(℃),WN34 CH4(℃),WN34 CH5(℃),WN34 CH6(℃),WN34 CH7(℃),WN34 CH8(℃),   LDS_Air CH1(mm),                               LDS_Air CH2(mm),                               LDS_Air CH3(mm),                               LDS_Air CH4(mm),
 			// Time,          CH1 Temperature(℃),CH1 Dew point(℃),CH1 HeatIndex(℃),CH1 Humidity(%), CH2 Temperature(℃),CH2 Dew point(℃),CH2 HeatIndex(℃),CH2 Humidity(%), CH3 Temperature(℃),CH3 Dew point(℃),CH3 HeatIndex(℃),CH3 Humidity(%),CH4 Temperature(℃), CH4 Dew point(℃),CH4 HeatIndex(℃),CH4 Humidity(%),CH5 Temperature(℃), CH5 Dew point(℃),CH5 HeatIndex(℃),CH5 Humidity(%),CH6 Temperature(℃), CH6 Dew point(℃),CH6 HeatIndex(℃),CH6 Humidity(%),CH7 Temperature(℃),CH7 Dew point(℃),CH7 HeatIndex(℃),CH7 Humidity(%),CH8 Temperature(℃),CH8 Dew point(℃), CH8 HeatIndex(℃),CH8 Humidity(%), WH35 CH1hum(%),WH35 CH2hum(%),WH35 CH3hum(%),WH35 CH4hum(%),WH35 CH5hum(%),WH35 CH6hum(%),WH35 CH7hum(%),WH35 CH8hum(%),             Thunder count,Thunder distance(km),AQIN Temperature(℃),AQIN Humidity(%),AQIN CO2(ppm),AQIN PM2.5(ug/m3),AQIN PM10(ug/m3),AQIN PM1.0(ug/m3),AQIN PM4.0(ug/m3),SoilMoisture CH1(%),                                                      SoilMoisture CH2(%),                                                      SoilMoisture CH3(%),                                                      SoilMoisture CH4(%),                                                      SoilMoisture CH5(%),                                                      SoilMoisture CH6(%),                                                      SoilMoisture CH7(%),                                                      SoilMoisture CH8(%),                                                      SoilMoisture CH9(%),                                                      SoilMoisture CH10(%),                                                         SoilMoisture CH11(%),                                                         SoilMoisture CH12(%),                                                         SoilMoisture CH13(%),                                                         SoilMoisture CH14(%),                                                         SoilMoisture CH15(%),                                                         SoilMoisture CH16(%),                                                         Water CH1,Water CH2,Water CH3,Water CH4,Pm2.5 CH1(ug/m3),Pm2.5 CH2(ug/m3),Pm2.5 CH3(ug/m3),Pm2.5 CH4(ug/m3),WN34 CH1(℃),WN34 CH2(℃),WN34 CH3(℃),WN34 CH4(℃),WN34 CH5(℃),WN34 CH6(℃),WN34 CH7(℃),WN34 CH8(℃),   LDS_Air CH1(mm),LDS_Depth CH1(mm),LDS_Heat CH1,LDS_Air CH2(mm),LDS_Depth CH2(mm),LDS_Heat CH2,LDS_Air CH3(mm),LDS_Depth CH3(mm),LDS_Heat CH3,LDS_Air CH4(mm),LDS_Depth CH4(mm),LDS_Heat CH4,
@@ -346,7 +348,7 @@ namespace CumulusMX.Stations
 			{
 				// invalid header
 				cumulus.LogErrorMessage("EcowittExtraLogFile.HeaderParser: Invalid header in file = " + header);
-				return;
+				return false;
 			}
 
 			// create a fields index
@@ -412,8 +414,9 @@ namespace CumulusMX.Stations
 					_ => LightningDist.km
 				};
 			}
-		}
 
+			return true;
+		}
 
 		private enum TempUnits
 		{