+++ CHANGELOG.md
@@ -22,6 +22,9 @@ Alternatively view it [online on GitHub](https://github.com/cumulusmx/CumulusMX/
 - Fix a major logic error when applying extra sensor data to the main station - affects most stations
 - Web tags are now case insensitive, as are tag parameter keys. Simple parameter values are also case insensitive. Parameter values for date formats etc are obviously still case sensitive
 - Fix MySQL error handling to prevent buffering of statements with syntax errors
+- Add support for BGT and WBGT to Ecowitt HTTP Local API, HTTP (Ecowitt), and the JSON stations
+	- New web tags `<#BlackGlobeTemp>` and `<#WetBulbGlobeTemp>`
+	- Two new fields added to the monthly log files and the monthly MySQL table to support these new measurements
 
 ### Notes
 
@@ -48,6 +51,10 @@ Alternatively view it [online on GitHub](https://github.com/cumulusmx/CumulusMX/
 - New Python script to upload monthly log files and the day file to MySQL - `/MXutils/ImportCumulusFile.py`
 - Adds logging of debug snow data via the Program Settings > Logging Options - Logs to `/data/debug_snowLog[sensornumber].txt`
 - Ecowitt HTTP Custom Server auto-configuration for main and extra stations now tries the HTTP Local API to access the station in addition to the TCP API
+- Add support for BGT and WBGT to Ecowitt HTTP Local API, HTTP (Ecowitt), and the JSON stations
+	- New web tags `<#BlackGlobeTemp>` and `<#WetBulbGlobeTemp>`
+	- Two new fields added to the monthly log files and the monthly MySQL table to support these new measurements
+- Fix ecowitt.net historic data download of PM measurements
 
 ### Changed